// src/types/enums.ts
export enum AccountStatus {
	NOT_VERIFIED = 'NOT_VERIFIED',
	VERIFIED = 'VERIFIED',
	EXPIRED = 'EXPIRED',
	TRIAL = 'TRIAL',
}

export enum BundleType {
	MONTHLY = 'MONTHLY',
	ONE_TIME = 'ONE_TIME',
}

export enum AdminRoles {
	SUPPORT = 'SUPPORT',
	SUPER = 'SUPER',
	OWNER = 'OWNER',
}